acrn: fix how to get memory below 4G
Fixes: e7c3a00353 ("Use ACRN E820 map to query system memory size")

In determine the range of memory to use during PEI, it assume the memory
below 4G or above 4G is continuous. However, if there's GPU DSM region
(which needs identical map), it would break the low memory into two
parts. In this case, we could let the OVMF see the lowmem part1.

Signed-off-by: Fei Li <fei1.li@intel.com>